refactor(admin/gigs): rename and refactor admin-gigs query state
